requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==65.3.0
django-ckeditor==6.7.2
django-cors-headers==4.6.0
django-js-asset==2.2.0
django-rest-knox==5.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==11.0.0
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.2
urllib3==2.2.3
